+++ docs/release_notes.rst
@@ -1,6 +1,12 @@
 Release Notes
 =============
 
+v6.2.0
+------
+
+* Added ``data_manifest`` option to `SwiftPath.download` and `SwiftPath.upload` for
+  generating data manifests and validation
+
 v6.1.0
 ------